// packages/cli/src/binding.d.ts
/* auto-generated by NAPI-RS */
/* eslint-disable */
export declare class AndroidClient {
  constructor(inputOptions: InputClientOptions)
  upload(): Promise<Array<CreateAppModal>>
}

export declare class CreateAppModal {
  appKey: string
  id: number
}

export interface BindingInputOptions {
  platform: PlatformEnum
  updateFile: string
}

export interface InputClientOptions {
  baseUrl: string
  name: string
  isDist: boolean
  userToken: string
}

export declare function loginCommand(params: LoginParams): Promise<void>

export interface LoginParams {
  password: string
  username: string
}

export declare enum PlatformEnum {
  Android = 0,
  Ios = 1,
  All = 2
}